move all friend notification body to one file
